[dev-3706] refactoring for testability
