+++ src/js/components/awardv2/shared/awardHistorySection/TablesSection.jsx
@@ -6,15 +6,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import TransactionsTableContainer from 'containers/awardV2/table/TransactionsTableContainer';
-import FederalAccountTableContainer from 'containers/awardV2/table/FederalAccountTableContainer';
-import { tabs, awardTypesWithSubawards } from 'dataMapping/awardsv2/awardHistorySection';
-
-import SubawardsContainer from '../../../../containers/awardV2/table/SubawardsContainer';
-import DetailsTabBar from '../../../award/details/DetailsTabBar';
-import ResultsTablePicker from '../../../search/table/ResultsTablePicker';
-import { getAwardHistoryCounts } from "../../../../helpers/awardHistoryHelper";
-
 const propTypes = {
     overview: PropTypes.object,
     activeTab: PropTypes.string,
@@ -27,28 +18,14 @@ export default class TablesSection extends React.Component {
         super(props);
 
         this.state = {
-            tableWidth: 0,
-            tabs: []
+            tableWidth: 0
         };
         this.countRequest = null;
         this.setTableWidth = this.setTableWidth.bind(this);
     }
 
     componentDidMount() {
-        // set the initial table width
-        this.setTableWidth();
-        // watch the window for size changes
-        window.addEventListener('resize', this.setTableWidth);
-        this.setTableTabsAndGetCounts();
-    }
 
-    componentDidUpdate(prevProps) {
-        // check award changed
-        if (this.props.overview.generatedId !== prevProps.overview.generatedId) {
-            // reset the tab
-            this.props.clickTab('transaction');
-            this.setTableTabsAndGetCounts();
-        }
     }
 
     componentWillUnmount() {
@@ -56,82 +33,8 @@ export default class TablesSection extends React.Component {
         window.removeEventListener('resize', this.setTableWidth);
     }
 
-    setTableTabsAndGetCounts(award = this.props.overview) {
-        if (this.countRequest) {
-            this.countRequest.cancel();
-        }
-
-        const tabsWithCounts = tabs(award.category)
-            .filter((tab) => {
-                if (tab.internal === 'subaward' && !awardTypesWithSubawards.includes(award.category)) {
-                    return false;
-                }
-                return true;
-            })
-            .map(async (tab) => {
-                const isIdv = (award.category === 'idv');
-                this.countRequest = getAwardHistoryCounts(tab.internal, award.generatedId, isIdv);
-                try {
-                    const { data } = await this.countRequest.promise;
-                    if (isIdv && tab.internal === 'federal_account') {
-                        // response object for idv federal account endpoint is { count: int }
-                        return { ...tab, count: data.count };
-                    }
-                    // response object for all other count endpoints are { [tab.internal + s] int }
-                    return { ...tab, count: data[`${tab.internal}s`] };
-                }
-                catch (error) {
-                    console.log(`Error fetching ${tab.internal} counts: ${error}`);
-                    return { ...tab, count: 'N/A' };
-                }
-            });
-
-        Promise.all(tabsWithCounts)
-            .then((result) => {
-                this.setState({ tabs: result }, this.setTableWidth);
-                this.countRequest = null;
-            });
-    }
-
-    setTableWidth() {
-        if (!this.tableWidthController) return;
-        const tableWidth = this.tableWidthController.clientWidth - 2;
-        this.setState({ tableWidth });
-    }
-
-    currentSection(
-        activeTab = this.props.activeTab,
-        overview = this.props.overview,
-        tableWidth = this.state.tableWidth,
-        awardId = this.props.awardId
-    ) {
-        switch (activeTab) {
-            case 'transaction':
-                return (
-                    <TransactionsTableContainer
-                        category={overview.category}
-                        tableWidth={tableWidth} />
-                );
-            case 'federal_account':
-                return (
-                    <FederalAccountTableContainer
-                        category={overview.category}
-                        tableWidth={tableWidth} />
-                );
-            case 'subaward':
-                return (
-                    <SubawardsContainer
-                        tableWidth={tableWidth}
-                        awardId={awardId} />
-                );
-            default:
-                return null;
-        }
-    }
-
     render() {
         const content = this.currentSection();
-        const tabOptions = this.state.tabs;
 
         if (tabOptions.length > 0) {
             return (
